updated_requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.62.0

# HuggingFace Hub
huggingface_hub>=0.15.0

# SAELens for GemmaScope SAEs
sae-lens>=0.1.0

# Network analysis
networkx>=2.6.0

# Additional utilities
requests>=2.25.0
Pillow>=8.3.0

# For correlation analysis
scikit-learn>=1.0.0

# For JSON handling
jsonschema>=3.2.0 
